# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexusRAG API",
    "version": "v1"
  },
  "paths": {
    "/v1/audio/{audio_id}.mp3": {
      "get": {
        "tags": [
          "audio"
        ],
        "summary": "Get Audio",
        "operationId": "get_audio_v1_audio__audio_id__mp3_get",
        "parameters": [
          {
            "name": "audio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Audio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "BAD_REQUEST",
                    "message": "Bad request"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_UNAUTHORIZED",
                    "message": "Missing or invalid bearer token"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "402": {
            "description": "Quota exceeded",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "QUOTA_EXCEEDED",
                    "message": "Monthly request quota exceeded",
                    "details": {
                      "period": "month",
                      "limit": 10000,
                      "used": 10000,
                      "remaining": 0
                    }
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_FORBIDDEN",
                    "message": "Insufficient role for this operation"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "Resource not found"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "IDEMPOTENCY_KEY_CONFLICT",
                    "message": "Idempotency-Key already used with different payload"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "REQUEST_VALIDATION_ERROR",
                    "message": "Validation error"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "RATE_LIMITED",
                    "message": "Rate limit exceeded",
                    "details": {
                      "scope": "api_key",
                      "route_class": "read",
                      "retry_after_ms": 1200
                    }
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INTERNAL_ERROR",
                    "message": "Internal server error"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "message": "Service unavailable"
                  },
                  "meta": {
                    "request_id": "req_example",
                    "api_version": "v1"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "error",
          "meta"
        ],
        "title": "ErrorEnvelope"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "v1"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}